+++ app/accounts/shemas.py
@@ -16,24 +16,5 @@ class ActualizarNombreCuenta(BaseModel):
 class ActualizarContrasenaCuenta(BaseModel):
     contrasena : str
 
-class Entradas(BaseModel):
-    origen : str
-    fecha : datetime
-    cantidad : float
-    detalle : Optional[str]
-    fk_cuenta_id : int
-    
-class ActualizarEntrada(BaseModel):
-    origen : Optional[str] = None
-    fecha : Optional[datetime] = None
-    cantidad : Optional[float] = None
-    detalle : Optional[str] = None
-    fk_cuenta_id : Optional[int] = None
 
-class Salidas(BaseModel):
-    categoria : str
-    descripcion : str
-    cantidad : float
-    fecha : datetime
-    detalle : Optional[str]
-    fk_cuenta_id : int
+
