+++ app/outflows/shemas.py
@@ -1,0 +1,11 @@
+from pydantic import BaseModel
+from typing import Optional
+from datetime import datetime
+
+class Salidas(BaseModel):
+    categoria : str
+    descripcion : str
+    cantidad : float
+    fecha : datetime
+    detalle : Optional[str]
+    fk_cuenta_id : int
